Rework homeFragment and ownStudyFragment
Because:
- Interaction Options on homeFragment was too much

Moving Filterlist from homeFragment to ownStudyFragment. Creating a dashboard on homeFragment. Replacing pieChart with progressbar which only displays completly booked vps.
Progressbar in ownStudyFragment will show current planned state.
