+++ app/src/main/java/com/example/vpmanager/views/CustomAlertReminder.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.view.Window;
 import android.widget.Button;
+import android.widget.TextView;
 
 import com.example.vpmanager.R;
 
@@ -13,11 +14,18 @@ public class CustomAlertReminder extends Dialog implements View.OnClickListener 
         public personalAccountFragment fragment;
         public Dialog d;
         public Button yes;
+        private TextView text;
 
 
         public CustomAlertReminder(personalAccountFragment a) {
                 super(a.getContext());
                 fragment = a;
+        }
+        public CustomAlertReminder(personalAccountFragment a, String alerText) {
+                super(a.getContext());
+
+                fragment = a;
+                text.setText(alerText);
         }
 
         @Override
@@ -25,7 +33,8 @@ public class CustomAlertReminder extends Dialog implements View.OnClickListener 
                 super.onCreate(savedInstanceState);
                 requestWindowFeature(Window.FEATURE_NO_TITLE);
                 setContentView(R.layout.custom_reminder);
-                yes = (Button) findViewById(R.id.reminder_dialog_ok);
+                text = findViewById(R.id.customAlertText);
+                yes = findViewById(R.id.reminder_dialog_ok);
                 yes.setOnClickListener(this);
         }
