+++ app/src/main/java/com/example/vpmanager/views/homeFragment.java
@@ -28,9 +28,6 @@ public class homeFragment extends Fragment {
     FirebaseAuth firebaseAuth;
 
 
-    private ViewPager viewPager;
-    private ViewPagerAdapter viewPagerAdapter;
-
     public homeFragment() {
     }
 
@@ -49,8 +46,6 @@ public class homeFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        navController = Navigation.findNavController(view);
-        createTabs(view);
     }
 
     @Override
@@ -58,48 +53,6 @@ public class homeFragment extends Fragment {
     {
         super.onStart();
         userLoggedIn();
-    }
-
-    //Parameter: current view to find components
-    //Return values:
-    //sets up the tablayout and adds and adapter for the content
-    private void createTabs(View view) {
-
-        viewPager = view.findViewById(R.id.view_pager_home);
-        TabLayout tabLayout = view.findViewById(R.id.tab_layout_home);
-
-        com.example.vpmanager.views.personalAccountFragment personalAccountFragment = new personalAccountFragment();
-        UpcomingAppointments appointments = new UpcomingAppointments();
-
-        tabLayout.addTab(tabLayout.newTab().setText("Persönliche Übersicht"));
-        tabLayout.addTab(tabLayout.newTab().setText("Termine"));
-
-        tabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-                viewPager.setCurrentItem(tab.getPosition());
-                tabLayout.selectTab(tab);
-            }
-
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
-
-            }
-
-            @Override
-            public void onTabReselected(TabLayout.Tab tab) {
-
-            }
-        });
-
-        viewPagerAdapter = new ViewPagerAdapter(getChildFragmentManager(), 0);//tabLayout.getTabCount());
-        viewPagerAdapter.addFragment(personalAccountFragment, "Persönliche Übersicht");
-        viewPagerAdapter.addFragment(appointments, "Termine");
-
-        viewPager.setAdapter(viewPagerAdapter);
-
-        viewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(tabLayout));
-        System.out.println("Tab Count: " + tabLayout.getTabCount());
     }
 
     //Parameter:
@@ -112,38 +65,5 @@ public class homeFragment extends Fragment {
         }
     }
 
-    //Parameter:
-    //Return values:
-    //defining adapter class for the tabLayout
-    private static class ViewPagerAdapter extends FragmentStatePagerAdapter
-    {
-        private final List<Fragment> fragments = new ArrayList<>();
-        private final List<String> fragmentNames = new ArrayList<>();
 
-        public ViewPagerAdapter(@NonNull FragmentManager fm, int behavior) {
-            super(fm, behavior);
-        }
-
-        public void addFragment(Fragment fragment, String title){
-            fragments.add(fragment);
-            fragmentNames.add(title);
-        }
-
-        @NonNull
-        @Override
-        public Fragment getItem(int position) {
-            return fragments.get(position);
-        }
-
-        @Override
-        public int getCount() {
-            return fragments.size();
-        }
-
-        @Nullable
-        @Override
-        public CharSequence getPageTitle(int position) {
-            return fragmentNames.get(position);
-        }
-    }
 }
