+++ app/src/main/res/layout/fragment_own_study.xml
@@ -1,52 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".views.ownStudyFragment"
+
     app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
-    <TextView
-        android:id="@+id/ownStudiesInfoText"
-        android:layout_width="250dp"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:text="Bisher hast du noch keine Studien erstellt. Wähle im Menü den Reiter Studien erstellen um Studien zu erstellen!"
-        android:textAlignment="center"
-        android:visibility="gone" />
-
-    <TextView
+    <com.google.android.material.tabs.TabLayout
+        android:id="@+id/tab_layout_ownStudies"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginStart="25dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="25dp"
-        android:text="Deine erstellten Studien"
-        android:textSize="20sp"
-        android:textStyle="bold" />
+        app:tabRippleColor="@color/green_light"
+        app:tabIndicatorColor="@color/green_Main"
+        app:tabSelectedTextColor="@color/white"
+        app:tabTextColor="@color/white"
+        android:clickable="true"
+        android:focusable="true">
 
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="3dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginTop="55dp"
-        android:layout_marginRight="20dp"
-        android:background="@color/color_two" />
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerViewOwnStudyFragment"
+    </com.google.android.material.tabs.TabLayout>
+
+
+
+    <androidx.viewpager.widget.ViewPager
+        android:id="@+id/view_pager_ownStudies"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="70dp"
-        android:dividerHeight="2dp" />
+        />
 
-    <!--
-    <ListView
-        android:id="@+id/listViewOwnStudyFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="70dp"
-        android:layout_weight="1"
-        android:dividerHeight="2dp" />
-        -->
-
-</RelativeLayout>
+</LinearLayout>
