+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -75,6 +75,9 @@
         <action
             android:id="@+id/action_ownStudyFragment_to_studyCreatorFragment"
             app:destination="@id/studyCreatorFragment" />
+        <action
+            android:id="@+id/action_ownStudyFragment_to_studyFragment"
+            app:destination="@id/studyFragment" />
     </fragment>
     <fragment
         android:id="@+id/studyFragment"
